Temporary work around for when loopring fails while we're placing an order.
This can still miss fill reports in the case where an order if placed, immedatly fills and then we fail it before we get the fill reports.  We'll do the full fix as a queue of these unknown orders that will either fail them or cancel them properly in the future.