Various updates to the dydx order tracking. This now, unfortunately, diverges us from the loopring connector's order tracking flow, but is required to deal with the dydx intricacies.
We now use the fill reports from the order stream on the websocket to get the fill price in-order to send out our fill events. There's a dance between the order status updates recieved and the fill reports recieved on when the events are issued to keep them in the correct order.

Added a unpad_price function to the token configuration.

Fixed a few bits of code based on the loopring connector feedback.